+++ CHANGELOG.md
@@ -19,6 +19,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - A default schema to be used when no schema is provided.
 - Child classes for `schema.Schema` that are specific to different serialization formats.
 
+### Fixed
+
+- Support for the latest LSST schema version (lsst.v7_1.alert). Note that this is the only LSST schema
+  version currently supported.
+
 ### Changed
 
 - Renamed `exceptions.PubSubInvalid` -> `exceptions.CloudConnectionError`, repurposed for more general use.